Add reusable CocktailRow component
